TEST: removed t_managed_exec.cc. Tests added to t_util.cc
