+++ cvmfs/util.h
@@ -167,7 +167,6 @@ bool HasPrefix(const std::string &str, const std::string &prefix,
                const bool ignore_case);
 bool HasSuffix(const std::string &str, const std::string &suffix,
                const bool ignore_case);
-bool IsNumeric(const std::string &str);
 
 std::vector<std::string> SplitString(const std::string &str,
                                      const char delim,
